backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
httpx==0.27.2
playwright==1.47.0
pydantic==2.9.2
pydantic-settings==2.5.2
anthropic==0.37.1
geopy==2.4.1
python-dotenv==1.0.1
tenacity==9.0.0
python-jose[cryptography]==3.3.0
stripe==11.1.0
svix==1.34.0
